+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+## [1.10.4]
 ### enhancements
 - clean weg items on load to remove duped items and apps/files that don't exist.
 - remove 1/2px thickness border on window manager border.